+++ filter.php
@@ -4444,6 +4444,29 @@ class filter_filtercodes extends moodle_text_filter {
                 }
             }
 
+            // Tag: {ifnottheme themename}...{/ifnottheme}.
+            // Description: Display content only if the current theme does not match the one specified.
+            // Parameters: The name of the directory in which the theme is located.
+            // Requires content between tags.
+            if (stripos($text, '{ifnottheme ') !== false) {
+                $theme = strtolower($PAGE->theme->name);
+                $re = '/{ifnottheme\s+(.*)\}(.*)\{\/ifnottheme\}/isuU';
+                $found = preg_match_all($re, $text, $matches);
+                if ($found > 0) {
+                    foreach ($matches[1] as $themename) {
+                        $key = '/{ifnottheme\s+' . $themename . '\}(.*)\{\/ifnottheme\}/isuU';
+                        if (strtolower($theme) != strtolower($themename)) {
+                            // Just remove the tags.
+                            $replace[$key] = '$1';
+                        } else {
+                            // Remove the ifnottheme strings.
+                            $replace[$key] = '';
+                        }
+                    }
+                }
+            }
+
+
             if (strpos($text, '{ifmin') !== false) { // If there are conditional ifmin tags.
                 // Tag: {ifminassistant}...{/ifminassistant}.
                 // Description: Display content only if user has the role of a non-editing teacher or higher.