+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [2.4.4] 2024-02-04 (dev)
+## [2.4.4] 2024-03-30 (dev)
 ### Added
 - New {iftheme}{/iftheme} tag.
 - New {ifnottheme}{/ifnottheme} tag.
 - New {alert border}{/alert} tag.
+- New links to edit Advanced theme settings and current theme settings to {menuthemes} tag.
 ### Updated
 - Fix-291: Escape tags now work reliably.
 - Copyright notice to include 2024.